+++ Pandora/Paradigm/Inventory/Ability/Replaceable.hs
@@ -3,5 +3,5 @@ module Pandora.Paradigm.Inventory.Ability.Replaceable where
 
 class Replaceable i where
 	type family Replacement i e r :: *
-	replace_:: Replacement i e r
+	replace:: Replacement i e r